Microchip XC8-v2.20_AVR_Sources
Downloaded from https://ww1.microchip.com/downloads/Secure/en/DeviceDoc/XC8-v2.20_AVR_Sources.zip

Release notes:

New device support
Support is available for the following AVR parts: ATTINY1624, AT-
TINY1626, and ATTINY1627.

Better best fit allocation
The best fit allocator (BFA) in the compiler has been improved so that
sections are allocated in an order permitting better optimization. The BFA now supports
named address spaces and better handles data initialization.

Improved procedural abstraction
The procedural abstraction optimizations are now performed
on more code sequences. Previous situations where this optimization might have increased
code size have been addressed by making the optimization code aware of the linker's
garbage collection process.

Absence of AVR Assembler
The AVR Assembler is no longer included with this distribution.

Changed DFP layout
The compiler now assumes a different layout used by DFPs (Device Family Packs).
This will mean that an older DFP might no work with this release, and older
compilers will not be able to use the latest DFPs.

Error with long commands (XC8-1983)
When using an AVR target, the compiler may have
stopped with a le not found error, if the command line was extremely large and contained
special characters such a quotes, backslashes, etc.

Unassigned rodata section (XC8-1920)
The AVR linker failed to assign memory for custom
rodata sections when building for avrxmega3 and avrtiny architectures, potentially produc-
ing memory overlap errors

+++ lto-plugin/lto-plugin.c
@@ -166,6 +166,8 @@ static unsigned int num_pass_through_items;
 
 static char debug;
 static char nop;
+static char keep_output;
+static char print_claimed_input;
 static char *resolution_file = NULL;
 
 /* The version of gold being used, or -1 if not gold.  The number is
@@ -523,6 +525,11 @@ cont:
 	= xrealloc (output_files, num_output_files * sizeof (char *));
       output_files[num_output_files - 1] = s;
       add_input_file (output_files[num_output_files - 1]);
+      if (print_claimed_input)
+        {
+          fprintf (stderr, "lto-plugin:added: %s\n", output_files[num_output_files - 1]);
+          fflush (stderr);
+        }
     }
 }
 
@@ -653,6 +660,11 @@ all_symbols_read_handler (void)
       struct plugin_file_info *info = &claimed_files[i];
 
       *lto_arg_ptr++ = info->name;
+      if (print_claimed_input)
+        {
+          fprintf (stderr, "lto-plugin:claimed: %s\n", info->name);
+          fflush (stderr);
+        }
     }
 
   for (i = 0; i < num_offload_files; i++)
@@ -704,10 +716,13 @@ cleanup_handler (void)
       check (t == 0, LDPL_FATAL, "could not unlink arguments file");
     }
 
-  for (i = 0; i < num_output_files; i++)
+  if (!keep_output)
     {
-      t = unlink (output_files[i]);
-      check (t == 0, LDPL_FATAL, "could not unlink output file");
+      for (i = 0; i < num_output_files; i++)
+        {
+          t = unlink (output_files[i]);
+          check (t == 0, LDPL_FATAL, "could not unlink output file");
+        }
     }
 
   free_2 ();
@@ -1018,6 +1033,10 @@ process_option (const char *option)
     debug = 1;
   else if (strcmp (option, "-nop") == 0)
     nop = 1;
+  else if (strcmp (option, "-keep-output") == 0)
+    keep_output = 1;
+  else if (strcmp (option, "-print-claimed-input") == 0)
+    print_claimed_input = 1;
   else if (!strncmp (option, "-pass-through=", strlen("-pass-through=")))
     {
       num_pass_through_items++;
